Added utils.py for Log into csv and Check Folder's valid
